java/product-project/task5/Test and fix of product and subclasses replaced doubles with big decimals in(product compuer smartphone and,Electronics) Added test case with to String to check the totalPrice

+++ src/com/company/Computer.java
@@ -1,10 +1,12 @@
 package com.company;
 
+import java.math.BigDecimal;
+
 public class Computer extends Product {
     private String processor;
     private int ram;
 
-    public Computer(int id, String name, double price, int stock) {
+    public Computer(int id, String name, BigDecimal price, int stock) {
         super(id, name, price, stock);
     }
