+++ src/com/company/Electronics.java
@@ -1,7 +1,9 @@
 package com.company;
 
+import java.math.BigDecimal;
+
 public class Electronics extends Product {
-    public Electronics(int id, String name, double price, int stock) {
+    public Electronics(int id, String name, BigDecimal price, int stock) {
         super(id, name, price, stock);
     }
